requirements.txt
openai>=1.0.0
openai-agents>=0.1.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=1.0.0
websockets>=12.0
